+++ api/core/repositories/__init__.py
@@ -2,5 +2,5 @@
 Repository implementations for data access.
 
 This package contains concrete implementations of the repository interfaces
-defined in the core.repository package.
+defined in the core.workflow.repository package.
 """
