+++ api/core/repositories/workflow_node_execution/__init__.py
@@ -2,7 +2,7 @@
 WorkflowNodeExecution repository implementations.
 """
 
-from repositories.workflow_node_execution.sqlalchemy_repository import SQLAlchemyWorkflowNodeExecutionRepository
+from core.repositories.workflow_node_execution.sqlalchemy_repository import SQLAlchemyWorkflowNodeExecutionRepository
 
 __all__ = [
     "SQLAlchemyWorkflowNodeExecutionRepository",
