+++ api/core/workflow/repository/__init__.py
@@ -6,8 +6,8 @@ for accessing and manipulating data, regardless of the underlying
 storage mechanism.
 """
 
-from core.repository.repository_factory import RepositoryFactory
-from core.repository.workflow_node_execution_repository import WorkflowNodeExecutionRepository
+from core.workflow.repository.repository_factory import RepositoryFactory
+from core.workflow.repository.workflow_node_execution_repository import WorkflowNodeExecutionRepository
 
 __all__ = [
     "RepositoryFactory",
